Функциональность Рекомендации. refactor: перенести метод getRecommendations и дополнительные private методы в FilmStorage
