+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
-import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -22,6 +21,4 @@ public interface UserStorage {
     List<User> getCommonFriends(Long id, Long friendId);
 
     void removeUserById(Long id);
-
-    List<Film> getRecommendations(Long id);
 }
